+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 
 * [MIC-6099] The payment button touch handling in the payment form, with a recurring payment
 
+### Added
+
+* [MIC-6063] Add TokenGenerator object
+
 ## [2.8.1] - 2022-08-05Z
 
 ### Fixed